[58489] feat: disable for PDF
