+++ src/main/java/sk/iway/iwcm/doc/ShowDoc.java
@@ -72,8 +72,6 @@ public class ShowDoc extends HttpServlet {
      */
     public static String ACTUAL_USER_HASH = null; //NOSONAR
 
-    private static final String MOVE_STYLE_TO_HEAD_KEY = "showDocMoveStyleToHead";
-
     /**
      * Skontroluje, ci parametre neobsahuju naznaky XSS, ak ano,
      * vrati redirect na URL bez skodlivych parametrov
@@ -306,11 +304,7 @@ private static String combineCss(String cssStyle)
 	{
 		request.setAttribute("templateDetails", temp);
 
-        if (temp.getMoveStyleToHead() != null && request.getAttribute(MOVE_STYLE_TO_HEAD_KEY) == null)
-        {
-            if (temp.getMoveStyleToHead().intValue() == 1) request.setAttribute(MOVE_STYLE_TO_HEAD_KEY, Boolean.TRUE);
-            else if (temp.getMoveStyleToHead().intValue() == 2) request.setAttribute(MOVE_STYLE_TO_HEAD_KEY, Boolean.FALSE);
-        }
+        StyleToHeadHelper.setRequestAttribute(temp, request);
 
 		long tempGroupId = temp.getTemplatesGroupId();
 		if (tempGroupId > 0)
@@ -1720,7 +1714,7 @@ private static String combineCss(String cssStyle)
     private void forwardWithBodyProcessing(String forward, HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
 
-        if (!isMoveStyleToHeadEnabled(request)) {
+        if (StyleToHeadHelper.isMoveStyleToHeadEnabled(request)==false) {
             // Feature disabled, use standard forward
             request.getRequestDispatcher(forward).forward(request, response);
             return;
@@ -1777,24 +1771,5 @@ private static String combineCss(String cssStyle)
 
         // Clear collected styles
         StyleToHeadHelper.clearCollectedStyles(request);
-    }
-
-    /**
-     * Returns true if move style to head feature is enabled for the current request, either via request attribute or global configuration.
-     * @param request The HTTP request
-     * @return true if the feature is enabled, false otherwise
-     */
-    public static boolean isMoveStyleToHeadEnabled(HttpServletRequest request) {
-
-        if ("true".equals(request.getParameter(SetCharacterEncodingFilter.PDF_PRINT_PARAM))) return false;
-        if (request.getParameter("isPdfVersion") != null) return false;
-
-        Boolean requestValue = (Boolean) request.getAttribute(MOVE_STYLE_TO_HEAD_KEY);
-
-        if (requestValue != null) {
-            return requestValue;
-        }
-
-        return Constants.getBoolean("showDocMoveStyleToHead"); //NOSONAR
     }
 }
